requirements.txt
requests
pandas
tqdm
python-dotenv
matplotlib
seaborn
scipy
numpy